refactor(fight-state): extract store and idle helper
